Refactor some VXL, OpenCV tests
Refactor the estimate_fundamental_matrix and estimate_homography tests
for VXL/OpenCV, which are virtually identical (besides obviously what
arrow they use, and a few tolerances; apparently, VXL produces closer to
"expected" results than OpenCV for some tests), into common sources in
order to avoid blatant code duplication. This also gains a VXL test,
since OpenCV had a test case that VXL did not.

Since the refactored tests use Google Test, this also ports these OpenCV
tests to Google Test.

Since it seemed strange to put the shared code in arrows/core, there is
now a new arrows/test directory for shared test code. The shared
landmark triangulation code has also been moved to here.
