+++ arrows/ocv/tests/test_estimate_fundamental_matrix.cxx
@@ -28,280 +28,35 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <test_common.h>
-#include <test_math.h>
-#include <test_scene.h>
-#include <test_random_point.h>
+#include <arrows/ocv/estimate_fundamental_matrix.h>
 
 #include <vital/plugin_loader/plugin_manager.h>
 
-#include <arrows/core/projected_track_set.h>
-#include <arrows/core/metrics.h>
-#include <arrows/core/epipolar_geometry.h>
+#include <gtest/gtest.h>
 
-#include <arrows/ocv/estimate_fundamental_matrix.h>
+using namespace kwiver::vital;
+using namespace kwiver::arrows;
 
-#include <Eigen/LU>
+using ocv::estimate_fundamental_matrix;
 
+static constexpr double ideal_tolerance = 1e-6;
+static constexpr double outlier_tolerance = 0.01;
 
-#define TEST_ARGS ()
-
-DECLARE_TEST_MAP();
-
-int
-main(int argc, char* argv[])
+// ----------------------------------------------------------------------------
+int main(int argc, char** argv)
 {
-  CHECK_ARGS(1);
+  ::testing::InitGoogleTest( &argc, argv );
 
   kwiver::vital::plugin_manager::instance().load_all_plugins();
 
-  testname_t const testname = argv[1];
-
-  RUN_TEST(testname);
+  return RUN_ALL_TESTS();
 }
 
-using namespace kwiver::vital;
-
-IMPLEMENT_TEST(create)
+// ----------------------------------------------------------------------------
+TEST(estimate_fundamental_matrix, create)
 {
-  using namespace kwiver::arrows;
-  algo::estimate_fundamental_matrix_sptr est_e = algo::estimate_fundamental_matrix::create("ocv");
-  if (!est_e)
-  {
-    TEST_ERROR("Unable to create ocv::estimate_fundamental_matrix by name");
-  }
+  EXPECT_NE( nullptr, algo::estimate_fundamental_matrix::create("ocv") );
 }
 
-
-// Print epipolar distance of pairs of points given a fundamental matrix
-void print_epipolar_distances(const kwiver::vital::matrix_3x3d& F,
-                              const std::vector<kwiver::vital::vector_2d> right_pts,
-                              const std::vector<kwiver::vital::vector_2d> left_pts)
-{
-  using namespace kwiver::arrows;
-  matrix_3x3d Ft = F.transpose();
-  for(unsigned i=0; i<right_pts.size(); ++i)
-  {
-    const vector_2d& pr = right_pts[i];
-    const vector_2d& pl = left_pts[i];
-    vector_3d vr(pr.x(), pr.y(), 1.0);
-    vector_3d vl(pl.x(), pl.y(), 1.0);
-    vector_3d lr = F * vr;
-    vector_3d ll = Ft * vl;
-    double sr = 1.0 / sqrt(lr.x()*lr.x() + lr.y()*lr.y());
-    double sl = 1.0 / sqrt(ll.x()*ll.x() + ll.y()*ll.y());
-    // sum of point to epipolar line distance in both images
-    double d = vr.dot(ll);
-    std::cout <<" dist right = "<<d*sr<<"  dist left = "<<d*sl << std::endl;
-  }
-}
-
-
-// test fundamental matrix estimation with ideal points
-IMPLEMENT_TEST(ideal_points)
-{
-  using namespace kwiver::arrows;
-  ocv::estimate_fundamental_matrix est_f;
-
-  // create landmarks at the random locations
-  landmark_map_sptr landmarks = kwiver::testing::init_landmarks(100);
-  landmarks = kwiver::testing::noisy_landmarks(landmarks, 1.0);
-
-  // create a camera sequence (elliptical path)
-  camera_map_sptr cameras = kwiver::testing::camera_seq();
-
-  // create tracks from the projections
-  track_set_sptr tracks = projected_tracks(landmarks, cameras);
-
-  const frame_id_t frame1 = 0;
-  const frame_id_t frame2 = 10;
-
-  camera_map::map_camera_t cams = cameras->cameras();
-  camera_sptr cam1 = cams[frame1];
-  camera_sptr cam2 = cams[frame2];
-  camera_intrinsics_sptr cal1 = cam1->intrinsics();
-  camera_intrinsics_sptr cal2 = cam2->intrinsics();
-
-  // compute the true fundamental matrix from the cameras
-  fundamental_matrix_sptr true_F = fundamental_matrix_from_cameras(*cam1, *cam2);
-
-  // extract coresponding image points
-  std::vector<track_sptr> trks = tracks->tracks();
-  std::vector<vector_2d> pts1, pts2;
-  for(unsigned int i=0; i<trks.size(); ++i)
-  {
-    auto fts1 = std::dynamic_pointer_cast<feature_track_state>(*trks[i]->find(frame1));
-    auto fts2 = std::dynamic_pointer_cast<feature_track_state>(*trks[i]->find(frame2));
-    pts1.push_back(fts1->feature->loc());
-    pts2.push_back(fts2->feature->loc());
-  }
-
-  // print the epipolar distances using this fundamental matrix
-  print_epipolar_distances(true_F->matrix(), pts1, pts2);
-
-  // compute the fundmental matrix from the corresponding points
-  std::vector<bool> inliers;
-  fundamental_matrix_sptr F_sptr = est_f.estimate(pts1, pts2,
-                                                  inliers, 1.5);
-  matrix_3x3d F = F_sptr->matrix();
-  // check for sign difference
-  if( true_F->matrix().cwiseProduct(F).sum() < 0.0 )
-  {
-    F *= -1;
-  }
-
-  // compare true and computed fundamental matrices
-  std::cout << "true F = "<<*true_F<<std::endl;
-  std::cout << "Estimated F = "<< F <<std::endl;
-  TEST_NEAR("Fundamental Matrix Estimate", F, true_F->matrix(), 1e-6);
-
-  unsigned num_inliers = static_cast<unsigned>(std::count(inliers.begin(),
-                                                          inliers.end(), true));
-  std::cout << "num inliers "<<num_inliers<<std::endl;
-  TEST_EQUAL("All points are inliers", num_inliers, pts1.size());
-}
-
-
-// test fundamental matrix estimation with noisy points
-IMPLEMENT_TEST(noisy_points)
-{
-  using namespace kwiver::arrows;
-  ocv::estimate_fundamental_matrix est_f;
-
-  // create landmarks at the random locations
-  landmark_map_sptr landmarks = kwiver::testing::init_landmarks(100);
-  landmarks = kwiver::testing::noisy_landmarks(landmarks, 1.0);
-
-  // create a camera sequence (elliptical path)
-  camera_map_sptr cameras = kwiver::testing::camera_seq();
-
-  // create tracks from the projections
-  feature_track_set_sptr tracks = projected_tracks(landmarks, cameras);
-
-  // add random noise to track image locations
-  tracks = kwiver::testing::noisy_tracks(tracks, 0.5);
-
-  const frame_id_t frame1 = 0;
-  const frame_id_t frame2 = 10;
-
-  camera_map::map_camera_t cams = cameras->cameras();
-  camera_sptr cam1 = cams[frame1];
-  camera_sptr cam2 = cams[frame2];
-  camera_intrinsics_sptr cal1 = cam1->intrinsics();
-  camera_intrinsics_sptr cal2 = cam2->intrinsics();
-
-  // compute the true fundamental matrix from the cameras
-  fundamental_matrix_sptr true_F = fundamental_matrix_from_cameras(*cam1, *cam2);
-
-  // extract coresponding image points
-  std::vector<track_sptr> trks = tracks->tracks();
-  std::vector<vector_2d> pts1, pts2;
-  for(unsigned int i=0; i<trks.size(); ++i)
-  {
-    auto fts1 = std::dynamic_pointer_cast<feature_track_state>(*trks[i]->find(frame1));
-    auto fts2 = std::dynamic_pointer_cast<feature_track_state>(*trks[i]->find(frame2));
-    pts1.push_back(fts1->feature->loc());
-    pts2.push_back(fts2->feature->loc());
-  }
-
-  print_epipolar_distances(true_F->matrix(), pts1, pts2);
-
-  // compute the fundamental matrix from the corresponding points
-  std::vector<bool> inliers;
-  fundamental_matrix_sptr F_sptr = est_f.estimate(pts1, pts2,
-                                                inliers, 1.5);
-  matrix_3x3d F = F_sptr->matrix();
-  // check for sign difference
-  if( true_F->matrix().cwiseProduct(F).sum() < 0.0 )
-  {
-    F *= -1;
-  }
-
-  // compare true and computed fundamental matrices
-  std::cout << "true F = "<<*true_F<<std::endl;
-  std::cout << "Estimated F = "<< F <<std::endl;
-  TEST_NEAR("Fundamental Matrix Estimate", F, true_F->matrix(), 0.01);
-
-  unsigned num_inliers = static_cast<unsigned>(std::count(inliers.begin(),
-                                                          inliers.end(), true));
-  std::cout << "num inliers "<<num_inliers<<std::endl;
-  bool enough_inliers = num_inliers > pts1.size() / 2;
-  TEST_EQUAL("Enough inliers", enough_inliers, true);
-}
-
-
-
-// test fundamental matrix estimation with outliers
-IMPLEMENT_TEST(outlier_points)
-{
-  using namespace kwiver::arrows;
-  ocv::estimate_fundamental_matrix est_f;
-
-  // create landmarks at the random locations
-  landmark_map_sptr landmarks = kwiver::testing::init_landmarks(100);
-  landmarks = kwiver::testing::noisy_landmarks(landmarks, 1.0);
-
-  // create a camera sequence (elliptical path)
-  camera_map_sptr cameras = kwiver::testing::camera_seq();
-
-  // create tracks from the projections
-  feature_track_set_sptr tracks = projected_tracks(landmarks, cameras);
-
-  // add random noise to track image locations
-  tracks = kwiver::testing::noisy_tracks(tracks, 0.5);
-
-  const frame_id_t frame1 = 0;
-  const frame_id_t frame2 = 10;
-
-  camera_map::map_camera_t cams = cameras->cameras();
-  camera_sptr cam1 = cams[frame1];
-  camera_sptr cam2 = cams[frame2];
-  camera_intrinsics_sptr cal1 = cam1->intrinsics();
-  camera_intrinsics_sptr cal2 = cam2->intrinsics();
-
-  // compute the true fundamental matrix from the cameras
-  fundamental_matrix_sptr true_F = fundamental_matrix_from_cameras(*cam1, *cam2);
-
-  // extract coresponding image points
-  std::vector<track_sptr> trks = tracks->tracks();
-  std::vector<vector_2d> pts1, pts2;
-  for(unsigned int i=0; i<trks.size(); ++i)
-  {
-    if (i % 3 == 0)
-    {
-      pts1.push_back(kwiver::testing::random_point2d(1000.0));
-      pts2.push_back(kwiver::testing::random_point2d(1000.0));
-    }
-    else
-    {
-      auto fts1 = std::dynamic_pointer_cast<feature_track_state>(*trks[i]->find(frame1));
-      auto fts2 = std::dynamic_pointer_cast<feature_track_state>(*trks[i]->find(frame2));
-      pts1.push_back(fts1->feature->loc());
-      pts2.push_back(fts2->feature->loc());
-    }
-  }
-
-  print_epipolar_distances(true_F->matrix(), pts1, pts2);
-
-  // compute the fundamental matrix from the corresponding points
-  std::vector<bool> inliers;
-  fundamental_matrix_sptr F_sptr = est_f.estimate(pts1, pts2,
-                                                  inliers, 1.5);
-  matrix_3x3d F = F_sptr->matrix();
-  // check for sign difference
-  if( true_F->matrix().cwiseProduct(F).sum() < 0.0 )
-  {
-    F *= -1;
-  }
-
-  // compare true and computed fundamental matrices
-  std::cout << "true F = "<<*true_F<<std::endl;
-  std::cout << "Estimated F = "<< F <<std::endl;
-  TEST_NEAR("Fundamental Matrix Estimate", F, true_F->matrix(), 0.01);
-
-  unsigned num_inliers = static_cast<unsigned>(std::count(inliers.begin(),
-                                                          inliers.end(), true));
-  std::cout << "num inliers "<<num_inliers<<" out of "<<pts1.size()<<" points."<<std::endl;
-  bool enough_inliers = num_inliers > pts1.size() / 3;
-  TEST_EQUAL("Enough inliers", enough_inliers, true);
-}
+// ----------------------------------------------------------------------------
+#include <arrows/tests/test_estimate_fundamental_matrix.h>
