+++ arrows/ocv/tests/test_estimate_homography.cxx
@@ -1,5 +1,5 @@
 /*ckwg +29
- * Copyright 2013-2016 by Kitware, Inc.
+ * Copyright 2013-2017 by Kitware, Inc.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -33,232 +33,45 @@
  * \brief test OCV homography estimation algorithm
  */
 
-#include <test_common.h>
+#include <test_eigen.h>
 #include <test_random_point.h>
-#include <vital/plugin_loader/plugin_manager.h>
 
 #include <arrows/ocv/estimate_homography.h>
 
-#include <opencv2/core/core.hpp>
+#include <vital/plugin_loader/plugin_manager.h>
+
+#include <gtest/gtest.h>
 
 using namespace kwiver::vital;
 using namespace kwiver::testing;
 using namespace kwiver::arrows;
 
-kwiver::vital::matrix_3x3d sample_homography()
+using ocv::estimate_homography;
+
+static constexpr double ideal_matrix_tolerance = 1e-4;
+static constexpr double ideal_norm_tolerance = 1e-4;
+
+static constexpr double noisy_matrix_tolerance = 0.1;
+static constexpr double noisy_norm_tolerance = 0.2;
+
+static constexpr double outlier_matrix_tolerance = 1e-4;
+static constexpr double outlier_norm_tolerance = 1e-4;
+
+// ----------------------------------------------------------------------------
+int main(int argc, char** argv)
 {
-  kwiver::vital::matrix_3x3d H;
-  H << 2.0, 0.0, -1.5,
-       0.0, 3.0, 5.0,
-       0.0, 0.0, 1.0;
-  return H;
-}
-
-#define TEST_ARGS ()
-
-DECLARE_TEST_MAP();
-
-int
-main(int argc, char* argv[])
-{
-  CHECK_ARGS(1);
-
-  testname_t const testname = argv[1];
+  ::testing::InitGoogleTest( &argc, argv );
 
   kwiver::vital::plugin_manager::instance().load_all_plugins();
 
-  RUN_TEST(testname);
+  return RUN_ALL_TESTS();
 }
 
-
-IMPLEMENT_TEST(create)
+// ----------------------------------------------------------------------------
+TEST(estimate_homography, create)
 {
-  algo::estimate_homography_sptr est_H = algo::estimate_homography::create("ocv");
-  if (!est_H)
-  {
-    TEST_ERROR("Unable to create ocv::estimate_homography by name");
-  }
+  EXPECT_NE( nullptr, algo::estimate_homography::create("ocv") );
 }
 
-
-IMPLEMENT_TEST(not_enough_points)
-{
-  ocv::estimate_homography estimator;
-
-  std::vector<vector_2d> pts1, pts2;
-  std::vector<bool> inliers;
-
-  homography_sptr H = estimator.estimate(pts1, pts2, inliers);
-  if ( H )
-  {
-    TEST_ERROR("Estimation with no points should produce a NULL homography");
-  }
-
-  pts1.push_back(vector_2d(0.0, 0.0));
-  pts1.push_back(vector_2d(2.0, 0.0));
-  pts1.push_back(vector_2d(0.0, 2.0));
-
-  pts2.push_back(vector_2d(1.0, 1.0));
-  pts2.push_back(vector_2d(4.0, 1.0));
-  pts2.push_back(vector_2d(1.0, 4.0));
-
-  H = estimator.estimate(pts1, pts2, inliers);
-  if ( H )
-  {
-    TEST_ERROR("Estimation with < 4 points should produce a NULL homography");
-  }
-}
-
-
-IMPLEMENT_TEST(four_points)
-{
-  ocv::estimate_homography estimator;
-
-  std::vector<vector_2d> pts1, pts2;
-  std::vector<bool> inliers;
-
-  pts1.push_back(vector_2d(0.0, 0.0));
-  pts1.push_back(vector_2d(2.0, 0.0));
-  pts1.push_back(vector_2d(0.0, 2.0));
-  pts1.push_back(vector_2d(2.0, 2.0));
-
-  matrix_3x3d true_H = sample_homography();
-
-  // transform pts1 to pts2 using true_H
-  for(unsigned i=0; i<pts1.size(); ++i)
-  {
-    vector_3d v = true_H * vector_3d(pts1[i].x(), pts1[i].y(), 1.0);
-    pts2.push_back(vector_2d(v.x()/v.z(), v.y()/v.z()));
-  }
-
-  homography_sptr H = estimator.estimate(pts1, pts2, inliers);
-  H = H->normalize();
-
-  double H_error = (true_H - H->matrix()).norm();
-  std::cout << "Homography estimation error: "<< H_error << std::endl;
-
-  std::cout << "H = "<<true_H <<std::endl;
-  TEST_NEAR("Frobenius norm between estimated and true homography",
-            H_error, 0.0, 1e-14);
-}
-
-
-IMPLEMENT_TEST(ideal_points)
-{
-  ocv::estimate_homography estimator;
-
-  matrix_3x3d true_H = sample_homography();
-
-  // create random points that perfectly correspond via true_H
-  std::vector<vector_2d> pts1, pts2;
-  for(unsigned i=0; i<100; ++i)
-  {
-    vector_2d v2 = random_point2d(1000.0) + vector_2d(500.0, 500.0);
-    pts1.push_back(v2);
-    vector_3d v3 = true_H * vector_3d(v2.x(), v2.y(), 1.0);
-    pts2.push_back(vector_2d(v3.x()/v3.z(), v3.y()/v3.z()));
-  }
-
-  std::vector<bool> inliers;
-  homography_sptr H = estimator.estimate(pts1, pts2, inliers);
-  H = H->normalize();
-
-  double H_error = (true_H - H->matrix()).norm();
-  std::cout << "Homography estimation error: "<< H_error << std::endl;
-  TEST_NEAR("Frobenius norm between estimated and true homography",
-            H_error, 0.0, 1e-4);
-
-  unsigned num_inliers = static_cast<unsigned>(std::count(inliers.begin(),
-                                                          inliers.end(), true));
-  std::cout << "num inliers "<<num_inliers<<std::endl;
-  TEST_EQUAL("All points are inliers", num_inliers, 100);
-}
-
-
-IMPLEMENT_TEST(noisy_points)
-{
-  ocv::estimate_homography estimator;
-
-  matrix_3x3d true_H = sample_homography();
-
-  // create random points + noise that approximately correspond via true_H
-  std::vector<vector_2d> pts1, pts2;
-  for(unsigned i=0; i<100; ++i)
-  {
-    vector_2d v2 = random_point2d(1000.0) + vector_2d(500.0, 500.0);
-    pts1.push_back(v2 + random_point2d(0.1));
-    vector_3d v3 = true_H * vector_3d(v2.x(), v2.y(), 1.0);
-    pts2.push_back(vector_2d(v3.x()/v3.z(), v3.y()/v3.z())
-                   + random_point2d(0.1));
-  }
-
-  std::vector<bool> inliers;
-  homography_sptr H = estimator.estimate(pts1, pts2, inliers);
-  H = H->normalize();
-
-  double H_error = (true_H - H->matrix()).norm();
-  std::cout << "Homography estimation error: "<< H_error << std::endl;
-  TEST_NEAR("Frobenius norm between estimated and true homography",
-            H_error, 0.0, 0.2);
-
-  unsigned num_inliers = static_cast<unsigned>(std::count(inliers.begin(),
-                                                          inliers.end(), true));
-  std::cout << "num inliers "<<num_inliers<<std::endl;
-  if (num_inliers < 75)
-  {
-    TEST_ERROR("Fewer than 75% of points detected as inliers");
-  }
-}
-
-
-IMPLEMENT_TEST(outlier_points)
-{
-  ocv::estimate_homography estimator;
-
-  matrix_3x3d true_H = sample_homography();
-
-  // create random points + noise that approximately correspond via true_H
-  std::vector<vector_2d> pts1, pts2;
-  std::vector<bool> true_inliers;
-  for(unsigned i=0; i<100; ++i)
-  {
-    vector_2d v2 = random_point2d(1000.0) + vector_2d(500.0, 500.0);
-    pts1.push_back(v2);
-    vector_3d v3 = true_H * vector_3d(v2.x(), v2.y(), 1.0);
-    pts2.push_back(vector_2d(v3.x()/v3.z(), v3.y()/v3.z()));
-    true_inliers.push_back(true);
-    if (i%3 == 0)
-    {
-      pts2.back() = random_point2d(1000.0) + vector_2d(500.0, 500.0);
-      true_inliers.back() = false;
-    }
-  }
-
-  std::vector<bool> inliers;
-  homography_sptr H = estimator.estimate(pts1, pts2, inliers);
-  H = H->normalize();
-
-  double H_error = (true_H - H->matrix()).norm();
-  std::cout << "Homography estimation error: "<< H_error << std::endl;
-  TEST_NEAR("Frobenius norm between estimated and true homography",
-            H_error, 0.0, 1e-4);
-
-  unsigned num_inliers = static_cast<unsigned>(std::count(inliers.begin(),
-                                                          inliers.end(), true));
-  std::cout << "num inliers "<<num_inliers<<std::endl;
-
-  unsigned correct_inliers = 0;
-  for( unsigned i=0; i<inliers.size(); ++i )
-  {
-    if (true_inliers[i] == inliers[i])
-    {
-      ++correct_inliers;
-    }
-  }
-  std::cout << "correct inliers "<<correct_inliers<<std::endl;
-
-  if (correct_inliers < 90)
-  {
-    TEST_ERROR("Fewer than 90% of points have correct inlier status");
-  }
-}
+// ----------------------------------------------------------------------------
+#include <arrows/tests/test_estimate_homography.h>
