+++ arrows/tests/README
@@ -1,0 +1,2 @@
+This directory contains implementations of test cases that are shared between
+multiple algorithm implementations.
